+++ scripts/fleetpage/content.js
@@ -86,7 +86,7 @@ const fleetQueueStr = fleetQueue.reduce((carry, el) => carry + cleanText(el.inne
 
 document.querySelector('#contentBox > .header')
     .insertAdjacentHTML('afterbegin', `
-        <span class="right copy-hint content-copy-handler">Click to copy to clipboard</span>
+        <span class="right copy-hint cursor-pointer">Click to copy to clipboard</span>
     `);
 document.querySelector('#contentBox .copy-hint')
     .addEventListener('click', e => {